x64: brgemm: jit_generator: eliminate implicit narrowing conversions
Fix -Wshorten-64-to-32/-Wimplicit-int-conversion/-Wfloat-conversion/
-Wimplicit-float-conversion/-Wsign-compare narrowing warnings in the
core JIT code-generation helpers shared by all brgemm kernels:

- jit_generator.hpp: wrap Xbyak immediate/predicate arguments (vpinsrb,
  vpinsrd, vpextrd, vcmpps, shift amounts, etc.) passed through the
  uni_v* wrapper functions with explicit static_cast<uint8_t>/
  Xbyak::uint8 at each call site.
- jit_generator.cpp: cast the vpinsrw immediate argument to uint8_t.

+++ src/cpu/x64/jit_generator.cpp
@@ -85,7 +85,8 @@ void jit_generator_t::transpose(const Xbyak::Reg64 &reg_src,
             // Load the remaining xf16 values one by one.
             for (int i = 0; i < rem; i++) {
                 vpinsrw(xmm_tmp, xmm_tmp,
-                        ptr[reg_src + r * src_stride + (c + i) * dt_size], i);
+                        ptr[reg_src + r * src_stride + (c + i) * dt_size],
+                        static_cast<uint8_t>(i));
             }
         }
